posix: preserve lower-risk output failures

+++ cmds/cp/atime_linux.go
@@ -12,7 +12,7 @@ import (
 // representation is unavailable; callers must not substitute another time.
 func atime(fi os.FileInfo) (time.Time, bool) {
 	if st, ok := fi.Sys().(*syscall.Stat_t); ok {
-		return time.Unix(st.Atim.Sec, st.Atim.Nsec), true
+		return time.Unix(int64(st.Atim.Sec), int64(st.Atim.Nsec)), true
 	}
 	return time.Time{}, false
 }
